fix(emulation): isolate forked state and bound assembly loading
"Isolated" forks shared one mutable `RuntimeState`, AppDomain and synthetic-method
map while running concurrently on rayon. `ThreadContext::fork` now deep-copies
that surface. Two exceptions are deliberate: `synthetic_method_counter` stays
shared, because private counters would let two forks mint the same token for
different bodies, and `native_functions` stays shared so a pointer captured in
one fork is interpretable in another.

`Assembly.Load(byte[])` grew two unbounded host-side collections and re-parsed
attacker metadata per call. Two new limits bound it: `max_loaded_assemblies` for
retention and `max_loaded_assembly_bytes` for a single payload before it reaches
the parser. Runtime-loaded assemblies parse with minimal validation -- they are
loaded for cross-assembly resolution, not accepted as trustworthy.

`assembly_index` widens from `u8` to `u32`, removing the crate's only truncation
escape. The PE loader cross-checks `SizeOfImage` against the section extents
instead of allocating it verbatim. `CilFlavor::FnPtr` is boxed, taking `EmValue`
from 200 to 104 bytes on x86-64, pinned by a static assertion.

+++ dotscope/src/emulation/engine/resolution.rs
@@ -38,7 +38,7 @@ pub enum CallResolution {
         /// Whether the caller expects a return value on the stack.
         expects_return: bool,
         /// Assembly index for dynamically loaded assemblies (`None` = primary).
-        assembly_index: Option<u8>,
+        assembly_index: Option<u32>,
         /// Method-level generic type arguments (`!!0`, `!!1`, ...) from MethodSpec.
         method_type_args: Option<Vec<Token>>,
     },
@@ -81,7 +81,7 @@ pub enum CallResolution {
         /// are wrapped in `TargetInvocationException`.
         is_reflection_invoke: bool,
         /// Assembly index for cross-assembly redirects (`None` = primary).
-        assembly_index: Option<u8>,
+        assembly_index: Option<u32>,
         /// Method-level generic type arguments (`!!0`, `!!1`, ...) from MethodSpec.
         method_type_args: Option<Vec<Token>>,
     },
